requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
deepagents>=0.0.10
langgraph>=0.4.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
tavily-python>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
sse-starlette>=2.0.0
